Add test case and refactor code
