+++ lumberjack/apps/ffmpeg/utils.py
@@ -7,3 +7,10 @@ def mkdir(dirname: str) -> None:
         os.makedirs(dirname)
     except OSError as exc:
         logging.info(exc)
+
+
+def generate_file_name_from_format(format):
+    if format.lower() == "mp4":
+        return "video.mp4"
+    elif format.lower() == "hls":
+        return "video.m3u8"
